requirements.txt
streamlit
pandas
scikit-learn
numpy
nltk
torch
transformers
google-genai
openai